refactoring moved plot stuffs to rtplot
